backend/requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0
motor==3.7.1
pydantic==2.12.5
python-dotenv==1.2.1
bcrypt==5.0.0
